# Fix the log duplication issue for --log-file