@dims 2019-07-30
benchmark with "fix"

```
dims@bigbox:~/go/src/k8s.io/klog$ go test -benchmem -run=^$ k8s.io/klog -bench "^(BenchmarkLogs)$" -v -count=10
goos: linux
goarch: amd64
pkg: k8s.io/klog
BenchmarkLogs-20    	  200000	      8123 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      7999 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      7580 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      8165 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      7790 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      8584 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      7789 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      7952 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      7665 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	      8583 ns/op	     556 B/op	       6 allocs/op
PASS
ok  	k8s.io/klog	25.716s
```

benchmark without the "fix"
```
dims@bigbox:~/go/src/k8s.io/klog$ go test -benchmem -run=^$ k8s.io/klog -bench "^(BenchmarkLogs)$" -v -count=10
goos: linux
goarch: amd64
pkg: k8s.io/klog
BenchmarkLogs-20    	  100000	     10248 ns/op	     560 B/op	       6 allocs/op
BenchmarkLogs-20    	  100000	     10330 ns/op	     560 B/op	       6 allocs/op
BenchmarkLogs-20    	  100000	     10996 ns/op	     560 B/op	       6 allocs/op
BenchmarkLogs-20    	  100000	     10331 ns/op	     560 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	     10166 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  200000	     10168 ns/op	     556 B/op	       6 allocs/op
BenchmarkLogs-20    	  100000	     10920 ns/op	     560 B/op	       6 allocs/op
BenchmarkLogs-20    	  100000	     10079 ns/op	     560 B/op	       6 allocs/op
BenchmarkLogs-20    	  100000	     10913 ns/op	     560 B/op	       6 allocs/op
BenchmarkLogs-20    	  100000	     10162 ns/op	     560 B/op	       6 allocs/op
PASS
ok  	k8s.io/klog	24.772s
```