@yuwenma 2019-08-07
@dims  The PR lgtm. The only concern I have is that: Even if user provides --log-file flag, they should still be able to use --alsologtostderr flag to require standard output. This requires storing logs individually by severity level. Also, in such case users should be able to use --stderrthreshold to change the severity level of the standard output (e.g. if stderrthreshold is warning, all logs in warning and above warning will be printed in the stdout).

The above feature is no longer available after this PR goes in. Is this what we want? If so, can you add a docstring for alsologtostderr and stderrthreshold flags to explain this behavior? 
 