@yuwenma 2019-08-07
> @yuwenma I don't understand your comment - it looks like the stderr output happens before these changes? Or were you commenting on an older version of this PR?

yeah, i was referring to a old change.

/lgtm